Extract shared token validation logic to concern
